refactor(web): split AdminDashboard into per-tab components (#19)
Extract each tab/section of the 2500+ line AdminDashboard.tsx into its
own file under web/src/components/admin/. AdminDashboard.tsx is now a
thin shell (134 lines) that manages tab state and dispatches to the
appropriate sub-component.

New components:
- ChoresTab, ApprovalsTab, RewardsTab, PointsTab, ActivityTab,
  UsersTab, SettingsTab, AIChoreChecker
- Supporting: ScheduleManager, TriggerManager, ExportConfigSection,
  APITokensSection

All components share the existing AdminDashboard.module.css. Behavior,
types, and API calls are preserved exactly — pure refactor.
